backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
pydantic==2.7.1
python-dotenv==1.0.1
websockets==12.0
aiofiles==23.2.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
